+++ src/ComposableFieldTypes/ComposableFieldRadiobuttons.jsx
@@ -7,16 +7,16 @@ export default class ComposableFieldRadiobuttons extends React.Component {
   render() {
     const options = this.props.fieldSettings.data || [];
     return(
-      <div className="radio_buttons" ref="options">
+      <div className="radio_buttons">
         {options.map((option, index) => {
-          var key = parameterize(this.props.fieldSettings.name) + "__option-" + index + "__" + parameterize(option.label + "");
+          var key = parameterize(this.props.fieldSettings.name) + "__option-" + index + "__" + parameterize(option.value + "");
           return(
             <span className="radio" key={key}>
               <label>
                 <Field
                   component="input"
                   type="radio"
-                  value={option.label}
+                  value={option.value + ""}
                   {...this.props.helpers.generateFieldAttributes(this.props)}
                 />
                 <span className='form--enhanced--control'></span>